Update to Muzei API 3.4
Use the new `getCommandActions()` API to fix the 'Download Wallpaper' action on Android 10+ devices. By keeping the old APIs as well, we maintain backward compatibility with Muzei 3.3 users (although they'll still need to upgrade to Muzei 3.4 to get a working acction).

The `PixelArtWorker` was changed to be compatible with the Kotlin rewrite of the Muzei API.

The `MuzeiArtDocumentsProvider` was added to the manifest to allow users to select artwork from the Source via the default file picker / Files app.

Muzei API 3.4 requires Kotlin 1.3.72 and a `compileSdkVersion` of 30 (which then required we upgrade the Android Gradle Plugin to 4.0.1), so those were also updated along with what other requirements went along with that.